Update macOS_Update.sh
Integration of a Debug Mode

Adds the option to enable or disable a debug mode.
Update Configuration

Via the configuration profile, it is now possible to specify how updates should be installed:
Latest version based on device eligibility
Latest major version
Latest minor version
Specific version
Critical Update Check

Checks the macOS version via sofa.macadmins.io to determine whether it is a critical update.
Display and Evaluation

UniqueCVEsCount
ActivelyExploitedCVEs
Deferral Rules

Defines the number of deferrals based on the ActivelyExploitedCVEs:
Non-critical updates: 5 days
Critical updates: 2 days
Pre-check Requirements

The macOS version is 14 (Sonoma) or later.
The Jamf Pro URL on the client can be identified.
Jamf Pro is reachable.
A bootstrap token has been generated by the client and submitted to Jamf Pro.
Jamf Pro Server Checks

Determine Device ID
Check if the bootstrap token is stored in Jamf Pro
If no token is stored, or if it has not been set by the client, the user is prompted to enter their password to generate the bootstrap token and send it to Jamf Pro.
(This is a one-time process. Once a token is created and stored, it will automatically be used to authorize macOS updates.)
Declarative Device Management

Check whether the client has declarativeDeviceManagementEnabled status:
If enabled, verify all existing plans for the client to see if there is an active plan.
If an active plan exists, evaluate it.
If no plan exists, create a new one.