Update macOS_Updates_DDM_Profile.json
new values ‘dontRunUpdatesOnDays’ it is possible to define days of the week on which the macOS update must not start.  Days on which the update may not be executed must be set to ‘true’ in the profile. 
The new value ‘runUpdatesOnTime’ can be used to specify the time at which the update is to be executed. If no hour or minutes are set, the current time of the macOS is used and defined in the plan. If only one value is set via the profile, the current variable is read from the macOS.