Update macOS_Update.sh
Implementation of Checks:
Whether the script is executed as root
Whether the Setup Assistant has been completed
Whether a user is logged in
Enhancement of the Jamf Pro availability check:
If the Jamf Pro URL cannot be reached via curl, the check is enforced via the Jamf Pro binary.
If the check via the binary is successful, the URL from the Jamf binary will be used.
Ensure the script runs with caffeinate
Check whether the device is not in screensaver mode
UI Adjustments:
The icon can now be defined via the configuration profile. If nothing is set in the profile, the icon defined in the script will be used.
To provide a more company-specific user interface, an overlay icon can now be defined over the main icon.
Either a custom overlay icon can be set, or the one already used in Self Service can be applied.
Bug Fixes:
Spelling corrections:
Funktion → Function
Hymand → Human
Fehlen → Missing
The MaxMessageTime variable was not correctly read from the profile – function adjusted.
The content of the variable latest_updateVersion_local_plist was sometimes displayed incorrectly (MSU_UPDATE_23H417 patch_15.3.1_minor).
The evaluation has been adjusted so that only the version number is now displayed: 15.3.1.