Update macOS_Updates_DDM_Profile.json
Text adjustments for multiple languages for SwiftDialog:
All texts in the dialog window in the InfoBar and the window for BootstrapToken are now transferred via the configuration profile. 
If the profile has not been set, English is set as the default text.
In the configuration profile, you can either choose between the texts or set your own texts or edit the texts. 

Now includes the option for setting silent mode